pyproject.toml
[tool.poetry]
name = "inicial-clientes"
version = "0.1.0"
description = ""
authors = ["Juan Dionizio <juan.dionzio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.0.3"
asgiref = "3.8.1"
astroid = "2.4.2"
colorama = "0.4.6"
django-rest-framework = "0.1.0"
djangorestframework = "3.15.1"
isort = "4.3.21"
lazy-object-proxy = "1.4.3"
mccabe = "0.6.1"
pylint = "2.5.3"
pytz = "2020.1"
six = "1.15.0"
sqlparse = "0.4.4"
toml = "0.10.1"
tzdata = "2024.1"
validate-docbr = "1.10.0"
wrapt = "1.12.1"
django-filter = "^24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
